refactor. fruitStore의 hasEnough와 subtract 매개변수 타입 수정
